requirements.txt
python-telegram-bot==20.7
python-jobspy==1.1.15
pandas
python-dotenv
pydantic==2.5.0
pydantic-settings==2.1.0
APScheduler==3.10.4
typer==0.9.0
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
